Command fix / interactions blacklist added / adding sync tasks / adding ssl for API

+++ events/client/guildDelete.js
@@ -1,3 +1,5 @@
 module.exports = async (client, guild) => {
-	return await client.Database.deleteGuild(guild.id);
+	let guilds = await client.Database.fetchGuild(guild.id);
+	if (!guilds.blacklisted)
+		return await client.Database.deleteGuild(guild.id);
 };
